// heroesApp/src/app/heroes/interfaces/hero.ts
// Generated by https://quicktype.io

export interface Hero {
    id:                string;
    superhero?:        string;
    publisher?:        Publisher;
    alter_ego?:        string;
    first_appearance?: string;
    characters?:       string;
    query?:            string;
    alt_img?:          string
}

export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}